Improve combination selections
