+++ toolbox.py
@@ -85,6 +85,35 @@ class SimpleCNN32Filter5Layers(torch.nn.Module):
         return x
 
 
+def write_result(filename, acc_ls):
+    """Writes results to specified text file"""
+    output = open(filename, "w")
+    for acc in acc_ls:
+        output.write(str(acc) + "\n")
+
+
+def combinations_45(iterable, r):
+    """Extracts 45 combinations from given list"""
+    pool = tuple(iterable)
+    n = len(pool)
+    if r > n:
+        return
+    indices = list(range(r))
+    yield tuple(pool[i] for i in indices)
+    count = 0
+    while count < 44:
+        count += 1
+        for i in reversed(range(r)):
+            if indices[i] != i + n - r:
+                break
+        else:
+            return
+        indices[i] += 1
+        for j in range(i + 1, r):
+            indices[j] = indices[j - 1] + 1
+        yield tuple(pool[i] for i in indices)
+
+
 def run_rf_image(
     model,
     train_images,
